FETCH'd and formatted news feed
