--- src/components/PostItem.js
@@ -1,5 +1,0 @@
-function PostItem(props) {
-  return <li>{props.author}</li>;
-}
-
-export default PostItem;
