+++ src/components/elements/Post.js
@@ -1,0 +1,7 @@
+import classes from './Post.module.css';
+
+function Post(props) {
+    return <div className={classes.post}>{props.children}</div>
+}
+
+export default Post;
