+++ src/components/elements/Post.module.css
@@ -1,0 +1,7 @@
+.post {
+    background-color: white;
+    border: 2px solid rgb(185, 185, 240);
+    display: flex;
+    flex-direction: column;
+    margin-bottom: 1rem;
+}
