+++ src/components/posts/PostItem.module.css
@@ -1,0 +1,28 @@
+.containerHeader {
+    align-items: center;
+    background-color: rgb(185, 185, 240);
+    display: flex;
+    height: 100%;
+    margin: 0;
+    padding-left: 2px;
+    width: 30rem;
+}
+
+.author {
+    color: white;
+    font-size: 1.2rem;
+}
+
+.content {
+    color: rgb(87, 87, 87);
+    padding: 2px;
+}
+
+.date {
+    display: flex;
+    font-size: 0.7rem;
+    max-height: 100%;
+    opacity: 0.5;
+    padding: 2px;
+    justify-content: flex-end;
+}
