+++ src/components/posts/PostList.module.css
@@ -1,0 +1,3 @@
+.list {
+    padding: 0;
+}
